created rank

+++ resources/views/book/book.blade.php
@@ -13,10 +13,10 @@
     <div class="book-top">
         <h1 class="book-title">図書詳細</h1>
         <div class="book-buttons">
-            <a class="book-button back" href="/book">戻る</a>
+            <a class="book-button back" href="/book/back">戻る</a>
             <a class="book-button edit" href="/book/edit/{{ $book['id'] }}">編集</a>
             <a class="book-button rate" href="/book/rate/{{ $book['id'] }}">評価</a>
-            <a class="book-button delete" href="/book/delete/{{ $book['id'] }}">削除</a>
+            <a class="book-button delete" href="/book/delete/{{ $book['id'] }}" onclick="return confirmDel();">削除</a>
         </div>
     </div>
